added extra panel for hud to display players. make sure to update game data! a new image is loaded!
this panel can currently not be deactivated (I'll work on this). the panel is displayed beside the mini map on lower resolutions and below the minimal on higher resolutions. It also displays defeated or lost players (due early disconnect). There are some odd behaivors at first game start and after loading a game.

git-svn-id: svn://maxr.org/MAX/trunk@1928 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b