Improved detection of a C++11/C++14 compatible compiler.
The checks are known not to work correctly with Apples version of clang, but since I do not have an Apple by hand at the moment I can not fix this.
The support for the Intel compiler is untested and very likely to be non-working since it depends on the cmake variable CMAKE_CXX_COMPILER_VERSION that seams to be very unreliable.